Develop

+++ src/Controller/Page/Post/PageController.php
@@ -177,11 +177,11 @@ class PageController extends AbstractInachisController
         $form->handleRequest($request);
 
         if ($form->isSubmitted()) {//} && $form->isValid()) {
-            if ($form->get('delete')->isClicked()) {
+            if ($form->has('delete') && $form->get('delete')->isClicked()) {
                 $this->entityManager->getRepository(Page::class)->remove($post);
                 $this->entityManager->getRepository(Revision::class)->deleteAndRecordByPage($post);
                 return $this->redirectToRoute(
-                    'app_dashboard_default',
+                    'incc_dashboard',
                     [],
                     Response::HTTP_PERMANENTLY_REDIRECT
                 );
@@ -250,7 +250,7 @@ class PageController extends AbstractInachisController
                 );
             }
 
-            if ($form->get('publish')->isClicked()) {
+            if ($form->has('publish') && $form->get('publish')->isClicked()) {
                 $post->setStatus(Page::PUBLISHED);
                 if (isset($revision)) {
                     $revision->setAction(RevisionRepository::PUBLISHED);
